+++ packages/components/src/color-scheme-switch/ColorSchemeSwitch.spec.tsx
@@ -2,7 +2,7 @@ import { beforeEach, describe, expect, it } from 'vitest'
 import { page, userEvent } from 'vitest/browser'
 import { composeStories } from '@storybook/react-vite'
 import * as stories from './ColorSchemeSwitch.stories'
-import { render } from 'vitest-browser-react'
+import { render } from '../../test-utils'
 
 const { Primary } = composeStories(stories)