+++ packages/datepicker-styles/CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.2.30 (2026-02-23)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.11.2
+
+## 0.2.29 (2026-02-23)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.11.1
+
+## 0.2.28 (2026-02-18)
+
+### 🧱 Updated Dependencies
+
+- Updated theme to 3.11.0
+
 ## 0.2.27 (2026-02-16)
 
 ### 🧱 Updated Dependencies